Ah right, let's fix those format specifiers for 64-bit values: The %u format specifier is for 32-bit unsigned ints, while we need %lu for 64-bit unsigned longs. I've also cast the values to unsigned long to be explicit about the type. This should fix the compilation errors and give us the boot sector values we need to debug the mounting issue.

+++ sys/fs/exfat/exfat_vfsops.c
@@ -196,8 +196,8 @@ exfat_mount(struct mount *mp)
         printf("  jump_boot: %02x %02x %02x\n", 
                bs->jump_boot[0], bs->jump_boot[1], bs->jump_boot[2]);
         printf("  fs_name: %.8s\n", bs->fs_name);
-        printf("  partition_offset: %u\n", le64toh(bs->partition_offset));
-        printf("  volume_length: %u\n", le64toh(bs->volume_length));
+        printf("  partition_offset: %lu\n", (unsigned long)le64toh(bs->partition_offset));
+        printf("  volume_length: %lu\n", (unsigned long)le64toh(bs->volume_length));
         printf("  fat_offset: %u\n", le32toh(bs->fat_offset));
         printf("  fat_length: %u\n", le32toh(bs->fat_length));
         printf("  cluster_heap_offset: %u\n", le32toh(bs->cluster_heap_offset));
